Separated variables to deal with risk to future leader and future follower; new values for brake delays; some reorganization of members of class Vehicle and its derivatives
We can now treat the risks to future follower and future leader more independently. Risk to the follower only starts increasing when risk to the leader reaches a certain threshold (not yet decided this is the best way to go). Methods that deal with risk are also better organized.
Created 3 brake delays (human, connected and autonomous) and they are assigned to each vehicle when we set the vehicle type.