Added connected vehicle following capacities to vehicles based on their type from VISSIM
Vehicles of type 120 are seen as connected. They have lower reaction times and use a control law that includes knowledge of the leader's acceleration.
Moreover, the estimates of the safe gaps for lane changing between connected vehicles are less conservative (no cooperative lane change yet)

Classes VehicleFollowingController and VelocityController are still stubs. They are supposed to contain most of the methods in LongitudinalVehicleController (and this will become more of a manager class).