Changes to gap selection and cooperation rule
If the ego vehicle has lane change intention and it has no origin lane leader, it will look for acceptable lane change gaps (keep moving forward) instead of trying to adjust to the nearest gap.
The ego vehicle will only cooperate for gap generation if it is at a minimum distance from the vehicle requesting cooperation.