Included lane change safe time headway in the origin lane controller and new reference velocity computation when generating gaps
To achieve lower risks, vehicles have to observe safe gaps to their origin lane leaders. This safe gap uses the lane changing max brake as parameters.
When generating gaps (either for itself or for an incoming vehicle), the ego vehicle assumes the minimum of its own speed and a fraction of the leader's speed as reference.